+++ sdk/nodejs/index.ts
@@ -25,9 +25,10 @@ export * from "./resource";
 // Export submodules individually.
 import * as asset from "./asset";
 import * as dynamic from "./dynamic";
+import * as iterable from "./iterable";
 import * as log from "./log";
 import * as runtime from "./runtime";
-export { asset, dynamic, log, runtime };
+export { asset, dynamic, iterable, log, runtime };
 
 // @pulumi is a deployment-only module.  If someone tries to capture it, and we fail for some reason
 // we want to give a good message about what the problem likely is.  Note that capturing a
